--- app/src/main/java/com/example/nanda/kotlinhub/listview_users.kt
@@ -1,12 +1,0 @@
-package com.example.nanda.kotlinhub
-
-import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-
-class listview_users : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_listview_users)
-    }
-}
