+++ app/src/main/res/layout/activity_profile_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ActivityProfilePage">
+    tools:context=".activities.ActivityProfilePage">
 
 
     <LinearLayout
