--- app/src/main/res/layout/activity_user_home.xml
@@ -1,143 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/black"
-    android:orientation="vertical"
-    tools:context=".activities.user_home">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_margin="0dp"
-        android:layout_weight="1"
-        android:background="@color/darkBlue"
-        android:orientation="horizontal"
-        android:padding="5dp">
-
-
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="-50dp"
-            android:layout_weight="1"
-            android:src="@drawable/user_img" />
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_marginLeft="8dp"
-            android:layout_weight="3"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/tv_name_field"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Name: Daniel Nan Da"
-                android:textSize="20sp" />
-
-            <TextView
-                android:id="@+id/tv_email_field"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Email: nanda@mail.com"
-                android:textSize="20sp"/>
-
-            <TextView
-                android:id="@+id/tv_score_field"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Progress: 24/50"
-                android:textSize="20sp"/>
-
-            <android.support.constraint.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-
-                <android.support.design.widget.FloatingActionButton
-                    android:id="@+id/floatingActionButton"
-                    android:layout_width="210dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginBottom="8dp"
-                    android:layout_marginEnd="8dp"
-                    android:layout_marginRight="8dp"
-                    android:clickable="true"
-                    android:cropToPadding="false"
-                    app:backgroundTint="@color/lightBlue"
-                    app:elevation="9dp"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:srcCompat="@drawable/edit_icon" />
-            </android.support.constraint.ConstraintLayout>
-
-        </LinearLayout>
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_margin="10dp"
-        android:layout_weight="2"
-        android:orientation="vertical">
-
-
-        <ProgressBar
-            android:id="@+id/progressBar2"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="5"
-            android:progress="80" />
-
-        <Button
-            android:id="@+id/btn_test"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:background="@color/gold"
-            android:text="Test(CM)" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:orientation="horizontal">
-            <Button
-                android:id="@+id/btn_home"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:background="@color/gold"
-                android:text="@string/btn_home" />
-
-            <Button
-                android:id="@+id/btn_lesson"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_marginLeft="1dp"
-                android:layout_marginRight="1dp"
-                android:layout_weight="1"
-                android:background="@color/gold"
-                android:text="@string/btn_lessons" />
-            <Button
-                android:id="@+id/btn_logout"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:background="@color/gold"
-                android:text="@string/btn_logout" />
-
-
-        </LinearLayout>
-
-
-    </LinearLayout>
-
-
-</LinearLayout>
