--- app/src/main/res/layout/content_user_profile.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".activities.user_profile"
-    tools:showIn="@layout/app_bar_user_profile">
-
-    <!--<include layout="@layout/activity_home_page" />-->
-        <FrameLayout
-            android:id="@+id/user_frame_layout"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-        </FrameLayout>
-
-
-</LinearLayout>
